Struct cs377_filesystem::myfs::MyFileSystem

pub struct MyFileSystem { /* private fields */ }

Implementations

impl MyFileSystem

pub fn new(disk_name: &str) -> MyFileSystem

Simply the equivalent of the constructor in Rust. Work with this filesystem on an existing file, we can create an instance:

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");

pub fn create_file(&mut self, filename: [u8; 8], size: u8) -> Result<(), String>

Creates a file in the filesystem with name capped at 8 bytes, and size can range from 0 to 8 This will check to see if creating the file is possible and will return an Err variant if not Usage:

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");
let filename: [u8; 8] = [102, 105, 108, 101, 49, 0, 0, 0]; //file1 as [u8; 8]
my_file_system.create_file(filename, 8);
my_file_system.ls();

This will output the following:

file1

pub fn delete_file(&mut self, filename: [u8; 8]) -> Result<(), String>

Deletes a file from the filesystem by marking the inode as unused and marking each block that was allocated for the file as unused if file exists. Otherwise, it returns the Err variant. Usage:

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");
let filename1: [u8; 8] = [102, 105, 108, 101, 49, 0, 0, 0]; //file1 as [u8; 8]
let filename2: [u8; 8] = [102, 105, 108, 101, 50, 0, 0, 0]; //file2 as [u8; 8]
my_file_system.create_file(filename1, 8);
my_file_system.create_file(filename2, 4);
my_file_system.delete_file(filename1);
my_file_system.ls();

This will output the following:

file2

pub fn ls(&mut self)

Prints all current files in the filesystem Usage:

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");
my_file_system.ls();

Will print nothing, as there are no files and just an empty filesystem. See other doctests for examples where something is printed by ls()

pub fn read( &mut self, filename: [u8; 8], block_num: u8 ) -> Result<[u8; 1024], String>

Reads block block_num out of file and returns Ok(contents) if it exists and returns Err otherwise. Usage:

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");
let filename: [u8; 8] = [102, 105, 108, 101, 49, 0, 0, 0]; //file1 as [u8; 8]
my_file_system.create_file(filename, 8);
println!("{:?}", my_file_system.read(filename, 7));

This will output one of the following assuming disk0 exists and the read was successful or not:

Ok("111....1111")

Or:

Err("Some error description")

pub fn write( &mut self, filename: [u8; 8], block_num: u8, write_buf: &[u8; 1024] ) -> Result<(), String>

Writes to block block_num of file and returns Ok(()) if successful and return Err otherwise. Usage:

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");
let filename: [u8; 8] = [102, 105, 108, 101, 49, 0, 0, 0]; //file1 as [u8; 8]
let my_new_data = [69u8; myfs::BLOCK_SIZE];
my_file_system.create_file(filename, 3);
my_file_system.write(filename, 2, &my_new_data);
println!("{:?}", my_file_system.read(filename, 2));

This will output one of the following assuming disk0 exists and the write was successful or not:

Ok("EEE....EEE")

Or

Err("Some error description")

pub fn close_disk(self)

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");
// do some stuff
my_file_system.close_disk();

And the following will not compile

use cs377_filesystem::myfs;
let mut my_file_system = myfs::MyFileSystem::new("disk0");
// do some stuff
my_file_system.close_disk();
my_file_system.ls()

With the error message

error[E0382]: borrow of moved value: `my_file_system`
--> src\lib.rs:247:1
    |
5   | let mut my_file_system = myfs::MyFileSystem::new("disk0");
    |     ------------------ move occurs because `my_file_system` has type `MyFileSystem`, which does not implement the `Copy` trait
6   | // do some stuff
7   | my_file_system.close_disk();
    |                ------------ `my_file_system` moved due to this method call
8   | my_file_system.ls()
    | ^^^^^^^^^^^^^^^^^^^ value borrowed here after move
    |
note: `MyFileSystem::close_disk` takes ownership of the receiver `self`, which moves `my_file_system`
--> C:\Repos\jack-champagne\cs377-final-project\src\lib.rs:254:27
    |
254 |         pub fn close_disk(self) {
    |                           ^^^^
error: aborting due to previous error